:hammer: Refactor: 파일 트리 재구성
