+++ src/components/ChatRoom/ChatDeclaration/ChatDeclarationStepDetail.tsx
